requirements.txt
numpy
scipy
matplotlib
pyusb
pywinusb
pywin32
pyudev
libusb1
